requirements.txt
# Core dependencies
crawl4ai>=0.4.3
python-dotenv>=0.19.0
pydantic>=1.9.0
beautifulsoup4>=4.10.0


# LLM API clients (install only what you need)
groq>=0.4.0  # For Groq models
anthropic>=0.5.0  # For Claude models
openai>=1.0.0  # For OpenAI models